requirements.txt
requests==2.22.0
python-telegram-bot==12.2.0
pytz==2019.3